+++ docs/js/menu-wc.js
@@ -1,41 +1,156 @@
 'use strict';
 
 
-
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
-
-customElements.define('compodoc-menu', function (_HTMLElement) {
-    _inherits(_class, _HTMLElement);
-
-    function _class() {
-        _classCallCheck(this, _class);
-
-        var _this = _possibleConstructorReturn(this, (_class.__proto__ || Object.getPrototypeOf(_class)).call(this));
-
-        _this.isNormalMode = _this.getAttribute('mode') === 'normal';
-        return _this;
+customElements.define('compodoc-menu', class extends HTMLElement {
+    constructor() {
+        super();
+        this.isNormalMode = this.getAttribute('mode') === 'normal';
     }
 
-    _createClass(_class, [{
-        key: 'connectedCallback',
-        value: function connectedCallback() {
-            this.render(this.isNormalMode);
-        }
-    }, {
-        key: 'render',
-        value: function render(isNormalMode) {
-            let tp = lithtml.html(
-'<nav>\n    <ul class="list">\n        <li class="title">\n            \n                <a href="index.html" data-type="index-link">wickeyappstore</a>\n            \n        </li>\n\n        <li class="divider"></li>\n        ' + (isNormalMode ? '<div id="book-search-input" role="search"><input type="text" placeholder="Type to search"></div>' : '') + '\n        <li class="chapter">\n            <a data-type="chapter-link" href="index.html"><span class="icon ion-ios-home"></span>Getting started</a>\n            <ul class="links">\n                \n                    <li class="link">\n                        <a href="overview.html" data-type="chapter-link">\n                            <span class="icon ion-ios-keypad"></span>Overview\n                        </a>\n                    </li>\n                    <li class="link">\n                        <a href="index.html" data-type="chapter-link">\n                            <span class="icon ion-ios-paper"></span>README\n                        </a>\n                    </li>\n                \n                \n                    <li class="link">\n                        \n                            <a href="changelog.html"\n                        \n                        data-type="chapter-link">\n                            <span class="icon ion-ios-paper"></span>CHANGELOG\n                        </a>\n                    </li>\n                \n                    <li class="link">\n                        \n                            <a href="license.html"\n                        \n                        data-type="chapter-link">\n                            <span class="icon ion-ios-paper"></span>LICENSE\n                        </a>\n                    </li>\n                \n                \n                    <li class="link">\n                        <a href="dependencies.html"\n                            data-type="chapter-link">\n                            <span class="icon ion-ios-list"></span>Dependencies\n                        </a>\n                    </li>\n                \n            </ul>\n        </li>\n        \n        \n        <li class="chapter modules">\n            <a data-type="chapter-link" href="modules.html">\n                <div class="menu-toggler linked" data-toggle="collapse"\n                    ' + (isNormalMode ? 'data-target="#modules-links"' : 'data-target="#xs-modules-links"') + '>\n                    <span class="icon ion-ios-archive"></span>\n                    <span class="link-name">Modules</span>\n                    <span class="icon ion-ios-arrow-down"></span>\n                </div>\n            </a>\n            <ul class="links collapse"\n            ' + (isNormalMode ? 'id="modules-links"' : 'id="xs-modules-links"') + '>\n                \n                    <li class="link">\n                        <a href="modules/WickeyAppStoreModule.html" data-type="entity-link">WickeyAppStoreModule</a>\n                        \n                            <li class="chapter inner">\n                                <div class="simple menu-toggler" data-toggle="collapse"\n                                    ' + (isNormalMode ? 'data-target="#components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"' : 'data-target="#xs-components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"') + '>\n                                    <span class="icon ion-md-cog"></span>\n                                    <span>Components</span>\n                                    <span class="icon ion-ios-arrow-down"></span>\n                                </div>\n                                <ul class="links collapse"\n                                    ' + (isNormalMode ? 'id="components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"' : 'id="xs-components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"') + '>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasAlert.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasAlert</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasLeaderboard.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasLeaderboard</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasPay.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasPay</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasProfile.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasProfile</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasReview.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasReview</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasSSO.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasSSO</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasShop.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasShop</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WasUp.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasUp</a>\n                                        </li>\n                                    \n                                        <li class="link">\n                                            <a href="components/WickeyAppStoreComponent.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">WickeyAppStoreComponent</a>\n                                        </li>\n                                    \n                                </ul>\n                            </li>\n                        \n                        \n                        \n                        \n                        \n                    </li>\n                \n            </ul>\n        </li>\n        \n        \n            \n        \n        \n        \n        \n        \n            \n                <li class="chapter">\n                    <div class="simple menu-toggler" data-toggle="collapse"\n                        ' + (isNormalMode ? 'data-target="#injectables-links"' : 'data-target="#xs-injectables-links"') + '>\n                        <span class="icon ion-md-arrow-round-down"></span>\n                        <span>Injectables</span>\n                        <span class="icon ion-ios-arrow-down"></span>\n                    </div>\n                    <ul class="links collapse"\n                    ' + (isNormalMode ? 'id="injectables-links"' : 'id="xs-injectables-links"') + '>\n                        \n                            <li class="link">\n                                <a href="injectables/UserService.html" data-type="entity-link">UserService</a>\n                            </li>\n                        \n                            <li class="link">\n                                <a href="injectables/WasDataService.html" data-type="entity-link">WasDataService</a>\n                            </li>\n                        \n                    </ul>\n                </li>\n            \n        \n        \n        \n        \n        <li class="chapter">\n            <div class="simple menu-toggler" data-toggle="collapse"\n                ' + (isNormalMode ? 'data-target="#interfaces-links"' : 'data-target="#xs-interfaces-links"') + '>\n                <span class="icon ion-md-information-circle-outline"></span>\n                <span>Interfaces</span>\n                <span class="icon ion-ios-arrow-down"></span>\n            </div>\n            <ul class="links collapse"\n            ' + (isNormalMode ? ' id="interfaces-links"' : 'id="xs-interfaces-links"') + '>\n                \n                    <li class="link">\n                        <a href="interfaces/Inapp.html" data-type="entity-link">Inapp</a>\n                    </li>\n                \n                    <li class="link">\n                        <a href="interfaces/User.html" data-type="entity-link">User</a>\n                    </li>\n                \n            </ul>\n        </li>\n        \n        \n        \n        \n        \n        <li class="chapter">\n            <a data-type="chapter-link" href="coverage.html"><span class="icon ion-ios-stats"></span>Documentation coverage</a>\n        </li>\n        \n        \n        \n    </ul>\n</nav>'
-);
-        this.innerHTML = tp.strings;
-        }
-    }]);
+    connectedCallback() {
+        this.render(this.isNormalMode);
+    }
 
-    return _class;
-}(HTMLElement));
+    render(isNormalMode) {
+        let tp = lithtml.html(`
+        <nav>
+            <ul class="list">
+                <li class="title">
+                    <a href="index.html" data-type="index-link">wickeyappstore</a>
+                </li>
+
+                <li class="divider"></li>
+                ${ isNormalMode ? `<div id="book-search-input" role="search"><input type="text" placeholder="Type to search"></div>` : '' }
+                <li class="chapter">
+                    <a data-type="chapter-link" href="index.html"><span class="icon ion-ios-home"></span>Getting started</a>
+                    <ul class="links">
+                        <li class="link">
+                            <a href="overview.html" data-type="chapter-link">
+                                <span class="icon ion-ios-keypad"></span>Overview
+                            </a>
+                        </li>
+                        <li class="link">
+                            <a href="index.html" data-type="chapter-link">
+                                <span class="icon ion-ios-paper"></span>README
+                            </a>
+                        </li>
+                        <li class="link">
+                            <a href="changelog.html"  data-type="chapter-link">
+                                <span class="icon ion-ios-paper"></span>CHANGELOG
+                            </a>
+                        </li>
+                        <li class="link">
+                            <a href="license.html"  data-type="chapter-link">
+                                <span class="icon ion-ios-paper"></span>LICENSE
+                            </a>
+                        </li>
+                        <li class="link">
+                            <a href="dependencies.html" data-type="chapter-link">
+                                <span class="icon ion-ios-list"></span>Dependencies
+                            </a>
+                        </li>
+                    </ul>
+                </li>
+                    <li class="chapter modules">
+                        <a data-type="chapter-link" href="modules.html">
+                            <div class="menu-toggler linked" data-toggle="collapse" ${ isNormalMode ?
+                                'data-target="#modules-links"' : 'data-target="#xs-modules-links"' }>
+                                <span class="icon ion-ios-archive"></span>
+                                <span class="link-name">Modules</span>
+                                <span class="icon ion-ios-arrow-down"></span>
+                            </div>
+                        </a>
+                        <ul class="links collapse" ${ isNormalMode ? 'id="modules-links"' : 'id="xs-modules-links"' }>
+                            <li class="link">
+                                <a href="modules/WickeyAppStoreModule.html" data-type="entity-link">WickeyAppStoreModule</a>
+                                    <li class="chapter inner">
+                                        <div class="simple menu-toggler" data-toggle="collapse" ${ isNormalMode ?
+                                            'data-target="#components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"' : 'data-target="#xs-components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"' }>
+                                            <span class="icon ion-md-cog"></span>
+                                            <span>Components</span>
+                                            <span class="icon ion-ios-arrow-down"></span>
+                                        </div>
+                                        <ul class="links collapse" ${ isNormalMode ? 'id="components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"' :
+                                            'id="xs-components-links-module-WickeyAppStoreModule-c983662ae4a65926ba02259903955614"' }>
+                                            <li class="link">
+                                                <a href="components/WasAlert.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasAlert</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WasLeaderboard.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasLeaderboard</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WasPay.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasPay</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WasProfile.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasProfile</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WasReview.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasReview</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WasSSO.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasSSO</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WasShop.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasShop</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WasUp.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WasUp</a>
+                                            </li>
+                                            <li class="link">
+                                                <a href="components/WickeyAppStoreComponent.html"
+                                                    data-type="entity-link" data-context="sub-entity" data-context-id="modules">WickeyAppStoreComponent</a>
+                                            </li>
+                                        </ul>
+                                    </li>
+                            </li>
+                </ul>
+                </li>
+                        <li class="chapter">
+                            <div class="simple menu-toggler" data-toggle="collapse" ${ isNormalMode ? 'data-target="#injectables-links"' :
+                                'data-target="#xs-injectables-links"' }>
+                                <span class="icon ion-md-arrow-round-down"></span>
+                                <span>Injectables</span>
+                                <span class="icon ion-ios-arrow-down"></span>
+                            </div>
+                            <ul class="links collapse" ${ isNormalMode ? 'id="injectables-links"' : 'id="xs-injectables-links"' }>
+                                <li class="link">
+                                    <a href="injectables/UserService.html" data-type="entity-link">UserService</a>
+                                </li>
+                                <li class="link">
+                                    <a href="injectables/WasDataService.html" data-type="entity-link">WasDataService</a>
+                                </li>
+                            </ul>
+                        </li>
+                    <li class="chapter">
+                        <div class="simple menu-toggler" data-toggle="collapse" ${ isNormalMode ? 'data-target="#interfaces-links"' :
+                            'data-target="#xs-interfaces-links"' }>
+                            <span class="icon ion-md-information-circle-outline"></span>
+                            <span>Interfaces</span>
+                            <span class="icon ion-ios-arrow-down"></span>
+                        </div>
+                        <ul class="links collapse" ${ isNormalMode ? ' id="interfaces-links"' : 'id="xs-interfaces-links"' }>
+                            <li class="link">
+                                <a href="interfaces/Inapp.html" data-type="entity-link">Inapp</a>
+                            </li>
+                            <li class="link">
+                                <a href="interfaces/User.html" data-type="entity-link">User</a>
+                            </li>
+                        </ul>
+                    </li>
+                    <li class="chapter">
+                        <a data-type="chapter-link" href="coverage.html"><span class="icon ion-ios-stats"></span>Documentation coverage</a>
+                    </li>
+            </ul>
+        </nav>
+        `);
+        this.innerHTML = tp.strings;
+    }
+});